+++ src/redi/cli/common.py
