+++ src/redi/cli/issue_command.py
@@ -7,7 +7,7 @@ from datetime import date
 from prompt_toolkit import prompt
 from prompt_toolkit.validation import Validator
 
-from redi.cli._common import (
+from redi.cli.common import (
     confirm_delete,
     inline_checkbox,
     inline_choice,
