+++ src/redi/cli/me_command.py
@@ -1,6 +1,6 @@
 import argparse
 
-from redi.cli._common import resolve_alias
+from redi.cli.common import resolve_alias
 from redi.i18n import messages
 from redi.api.me import read_my_account, update_my_account
