+++ src/redi/cli/role_command.py
@@ -1,6 +1,6 @@
 import argparse
 
-from redi.cli._common import resolve_alias
+from redi.cli.common import resolve_alias
 from redi.i18n import messages
 from redi.api.role import list_roles, read_role
