+++ src/redi/cli/wiki_command.py
@@ -3,7 +3,7 @@ import argparse
 from prompt_toolkit import prompt
 from prompt_toolkit.validation import ValidationError, Validator
 
-from redi.cli._common import confirm_delete, inline_choice, open_editor, resolve_alias
+from redi.cli.common import confirm_delete, inline_choice, open_editor, resolve_alias
 from redi.config import default_project_id, wiki_project_id
 from redi.i18n import messages
 from redi.api.wiki import (
